Refactor: 머렝 비즈니스 예외로 다시 변경
