+++ mureng-api/src/main/java/net/mureng/api/core/advice/DefaultControllerAdvice.java
@@ -2,7 +2,7 @@ package net.mureng.api.core.advice;
 
 import net.mureng.api.core.dto.ApiResult;
 import net.mureng.core.core.exception.*;
-import net.mureng.core.core.exception.business.MurengException;
+import net.mureng.core.core.exception.MurengException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
