+++ mureng-api/src/main/java/net/mureng/api/member/service/MemberBadgeService.java
@@ -5,9 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import net.mureng.core.badge.entity.BadgeAccomplished;
 import net.mureng.core.badge.entity.BadgeAccomplishedPK;
 import net.mureng.core.badge.repository.BadgeAccomplishedRepository;
-import net.mureng.core.core.exception.BadRequestException;
-import net.mureng.core.core.exception.ResourceNotFoundException;
-import net.mureng.core.core.exception.business.EntityNotFoundException;
+import net.mureng.core.core.exception.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
