+++ mureng-api/src/main/java/net/mureng/api/web/HomeController.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import net.mureng.api.core.dto.ApiResult;
-import net.mureng.core.core.exception.business.MurengException;
+import net.mureng.core.core.exception.MurengException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
