+++ mureng-api/src/test/java/net/mureng/api/reply/web/ReplyLikesControllerTest.java
@@ -39,7 +39,7 @@ public class ReplyLikesControllerTest extends AbstractControllerTest {
         mockMvc.perform(
                 post("/api/reply/{replyId}/reply-likes", 1)
                         .contentType(MediaType.APPLICATION_JSON)
-        ).andExpect(status().isBadRequest())
+        ).andExpect(status().isInternalServerError())
                 .andDo(print());
     }
 
@@ -61,7 +61,7 @@ public class ReplyLikesControllerTest extends AbstractControllerTest {
         mockMvc.perform(
                 delete("/api/reply/{replyId}/reply-likes", 2)
                         .contentType(MediaType.APPLICATION_JSON)
-        ).andExpect(status().isBadRequest())
+        ).andExpect(status().isInternalServerError())
                 .andDo(print());
     }
 }
