+++ mureng-core/src/main/java/net/mureng/core/badge/service/BadgeService.java
@@ -3,8 +3,7 @@ package net.mureng.core.badge.service;
 import lombok.RequiredArgsConstructor;
 import net.mureng.core.badge.entity.Badge;
 import net.mureng.core.badge.repository.BadgeRepository;
-import net.mureng.core.core.exception.ResourceNotFoundException;
-import net.mureng.core.core.exception.business.EntityNotFoundException;
+import net.mureng.core.core.exception.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
