+++ mureng-core/src/main/java/net/mureng/core/core/exception/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package net.mureng.core.core.exception;
 /**
  * 리소스 찾을 수 없음 예외 (404 Not Found)
  */
-public class ResourceNotFoundException extends RuntimeException {
+public class ResourceNotFoundException extends MurengException {
     public ResourceNotFoundException() {
     }
